v1.0.0/raspberry-pi/requirements.txt
RPi.GPIO==0.7.1
pynmea2==1.19.0
pyserial==3.5 